@@ -5,13 +5,10 @@ import { Cart } from "./cart.model";
 import { Order } from "./order.model";
 import { OrderRepository } from "./order.repository";
 import { RestDataSource } from "./rest.datasource";
-import { HttpClientModule } from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import { AuthService } from "./auth.service";
 
-@NgModule({
-  imports: [HttpClientModule],
-  providers: [ProductRepository, Cart, Order, OrderRepository,
-    { provide: StaticDataSource, useClass: RestDataSource },
-    RestDataSource, AuthService]
-})
+@NgModule({ imports: [], providers: [ProductRepository, Cart, Order, OrderRepository,
+        { provide: StaticDataSource, useClass: RestDataSource },
+        RestDataSource, AuthService, provideHttpClient(withInterceptorsFromDi())] })
 export class ModelModule { }
